fix scaling

+++ parser/test/mppt_simulator.py
@@ -221,8 +221,8 @@ class SimDevice:
             "vbat_max_mv": int(round(base_vbat + _rnd(0.2, 0.8, 3), 3) * 1000),
             "vpv_min_mv": int(round(max(0.0, base_vpv - _rnd(5.0, 10.0, 3)), 3) * 1000),
             "vpv_max_mv": int(base_vpv * 1000),
-            "ah_charge_mah": int(_rnd(10.0, 70.0, 1)),
-            "ah_load_mah": int(_rnd(8.0, 60.0, 1)),
+            "ah_charge_mah": _rnd(10.0, 70.0, 1),
+            "ah_load_mah": _rnd(8.0, 60.0, 1),
             "il_max_ma10": int(_rnd(0.5, 12.0, 2) * 10),
             "ipv_max_ma10": int(_rnd(1.0, 10.0, 2) * 10),
             "soc_pct": int(_rnd_int(5, 100)),
@@ -241,8 +241,8 @@ class SimDevice:
             "vbat_max_mv": int(_rnd(13.5, 14.2, 3) * 1000),
             "vpv_min_mv": int(_rnd(3.0, 7.0, 3) * 1000),
             "vpv_max_mv": int(_rnd(17.0, 22.0, 3) * 1000),
-            "ah_charge_mah": int(round(200.0 + season * 700.0 + _rnd(-50, 50, 1), 1)),
-            "ah_load_mah": int(round(150.0 + season * 500.0 + _rnd(-30, 30, 1), 1)),
+            "ah_charge_mah": round(200.0 + season * 700.0 + _rnd(-50, 50, 1), 1),
+            "ah_load_mah": round(150.0 + season * 500.0 + _rnd(-30, 30, 1), 1),
             "il_max_ma10": int(_rnd(1.0, 4.0, 2) * 10),
             "ipv_max_ma10": int(_rnd(2.0, 8.0, 2) * 10),
             "soc_pct": int(_rnd_int(20, 90)),
@@ -447,8 +447,8 @@ class SimDevice:
         entry.vbat_max_mv = row["vbat_max_mv"]
         entry.vpv_min_mv = row["vpv_min_mv"]
         entry.vpv_max_mv = row["vpv_max_mv"]
-        entry.ah_charge_mah = row["ah_charge_mah"]
-        entry.ah_load_mah = row["ah_load_mah"]
+        entry.ah_charge_mah = int(row["ah_charge_mah"])
+        entry.ah_load_mah = int(row["ah_load_mah"])
         entry.il_max_ma10 = row["il_max_ma10"]
         entry.ipv_max_ma10 = row["ipv_max_ma10"]
         entry.soc_pct = row["soc_pct"]
